util/python/requirements.txt
uuid
psycopg2-binary